Create BaseInput class Extend BaseInput with TMInput
