+++ DemoApp/DemoApp/MainPage.xaml
@@ -1,9 +1,15 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:Modus="clr-namespace:Trimble.Modus.Components.Controls.TMInput;assembly=Trimble.Modus.Components"
              Title="Controls Page"
-             x:Class="DemoApp.MainPage">
+             x:Class="DemoApp.MainPage"
+             xmlns:constants="clr-namespace:DemoApp.Constant">
+  <!--<VerticalStackLayout Spacing="10">
+        <Modus:TMInput IsReadOnly="False" Placeholder="Placeholder" HelperText="Helper" TitleText="Title" LeftIconSource="{x:Static constants:ImageConstants.PasswordIcon}" />
 
+    <Modus:MultiLineInput />
+  </VerticalStackLayout>-->
   <ListView ItemsSource="{Binding ControlNames}"
             ItemTapped="ItemTapped">
     <ListView.ItemTemplate>
