+++ DemoApp/DemoApp/Views/TMInputPage.xaml.cs
@@ -7,6 +7,7 @@ namespace DemoApp;
 public partial class TMInputPage : ContentPage
 {
     private InputPageViewModel _inputPageViewModel;
+
     public TMInputPage()
     {
         _inputPageViewModel = new InputPageViewModel();
