+++ Trimble.Modus.Components/Controls/TMInput/InputTemplateResourceDictionary.xaml.cs
@@ -1,0 +1,10 @@
+﻿namespace Trimble.Modus.Components.Controls.TMInput;
+
+public partial class InputTemplateResourceDictionary : ResourceDictionary
+{
+    public InputTemplateResourceDictionary()
+    {
+        InitializeComponent();
+    }
+}
+
