+++ Trimble.Modus.Components/Controls/TMInput/MultiLineInput.xaml
@@ -1,0 +1,17 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+  ControlTemplate="{StaticResource InputTemplate}"
+             VerticalOptions="FillAndExpand"
+             x:Class="Trimble.Modus.Components.Controls.TMInput.MultiLineInput">
+  <ContentView.Resources>
+        <ResourceDictionary>
+            <ResourceDictionary.MergedDictionaries>
+                <ResourceDictionary Source="InputTemplateResourceDictionary.xaml" />
+            </ResourceDictionary.MergedDictionaries>
+        </ResourceDictionary>
+    </ContentView.Resources>
+    <Editor Placeholder="Enter something"
+            AutoSize="TextChanges"></Editor>
+</ContentView>
+
