+++ Trimble.Modus.Components/Controls/TMInput/MultiLineInput.xaml.cs
@@ -1,0 +1,10 @@
+﻿namespace Trimble.Modus.Components.Controls.TMInput;
+
+public partial class MultiLineInput : ContentView
+{
+    public MultiLineInput()
+    {
+        InitializeComponent();
+    }
+}
+
